+++ L1Trigger/Phase2L1ParticleFlow/src/l1-converters/hgcalinputt_ref.cpp
@@ -1,4 +1,4 @@
-#include "hgcalinput_ref.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/l1-converters/hgcalinput_ref.h"
 
 l1ct::HgcalClusterDecoderEmulator::~HgcalClusterDecoderEmulator() {}
