+++ L1Trigger/Phase2L1ParticleFlow/src/l1-converters/tkinput_ref.cpp
@@ -1,13 +1,9 @@
-#include "tkinput_ref.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/l1-converters/tkinput_ref.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/dbgPrintf.h"
+
 #include <cmath>
 #include <cassert>
 #include <cstdio>
-
-#ifdef CMSSW_GIT_HASH
-#include "L1Trigger/Phase2L1ParticleFlow/interface/dbgPrintf.h"
-#else
-#include "../../../utils/dbgPrintf.h"
-#endif
 
 #ifdef CMSSW_GIT_HASH
 #include <cstdint>
