+++ L1Trigger/Phase2L1ParticleFlow/src/pf/pfalgo2hgc_ref.cpp
@@ -1,15 +1,10 @@
-#include "pfalgo2hgc_ref.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/pf/pfalgo2hgc_ref.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/dbgPrintf.h"
 
 #include <cmath>
 #include <cstdio>
 #include <algorithm>
 #include <memory>
-
-#ifdef CMSSW_GIT_HASH
-#include "L1Trigger/Phase2L1ParticleFlow/interface/dbgPrintf.h"
-#else
-#include "../../utils/dbgPrintf.h"
-#endif
 
 #ifdef CMSSW_GIT_HASH
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
