+++ L1Trigger/Phase2L1ParticleFlow/src/pf/pfalgo_common_ref.cpp
@@ -1,13 +1,8 @@
-#include "pfalgo_common_ref.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/pf/pfalgo_common_ref.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/dbgPrintf.h"
 
 #include <cmath>
 #include <cstdio>
-
-#ifdef CMSSW_GIT_HASH
-#include "L1Trigger/Phase2L1ParticleFlow/interface/dbgPrintf.h"
-#else
-#include "../../utils/dbgPrintf.h"
-#endif
 
 #ifdef CMSSW_GIT_HASH
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
