+++ L1Trigger/Phase2L1ParticleFlow/src/regionizer/multififo_regionizer_ref.cpp
@@ -1,10 +1,9 @@
-#include "multififo_regionizer_ref.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/regionizer/multififo_regionizer_ref.h"
 #include "L1Trigger/Phase2L1ParticleFlow/interface/egamma/pfeginput_ref.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/regionizer/multififo_regionizer_elements_ref.icc"
 
 #include <iostream>
 #include <memory>
-
-#include "multififo_regionizer_elements_ref.icc"
 
 #ifdef CMSSW_GIT_HASH
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
