+++ L1Trigger/Phase2L1ParticleFlow/src/regionizer/tdr_regionizer_ref.cpp
@@ -1,8 +1,7 @@
-#include "tdr_regionizer_ref.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/regionizer/tdr_regionizer_ref.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/regionizer/tdr_regionizer_elements_ref.icc"
 
 #include <iostream>
-
-#include "tdr_regionizer_elements_ref.icc"
 
 #ifdef CMSSW_GIT_HASH
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
